test(types): move app_context message coverage into message tests
