+++ packages/types/test/events/app.test-d.ts
@@ -1,5 +1,5 @@
 import { expectAssignable, expectError } from 'tsd';
-import type { AppContextChangedEvent, GenericMessageEvent, SlackEvent } from '../../src/index';
+import type { AppContextChangedEvent, SlackEvent } from '../../src/index';
 
 // a channel_id context entity
 const channelContextChangedEvent: AppContextChangedEvent = {
@@ -73,27 +73,3 @@ expectError<AppContextChangedEvent>({
   },
   event_ts: '1234567890.123456',
 });
-
-// the context is delivered inline on messages via the `app_context` field
-const messageWithAppContext: GenericMessageEvent = {
-  type: 'message',
-  subtype: undefined,
-  channel_type: 'im',
-  channel: 'D0123456789',
-  user: 'U0123456789',
-  ts: '1234567890.123456',
-  event_ts: '1234567890.123456',
-  text: 'hi',
-  app_context: {
-    entities: [
-      {
-        type: 'slack#/types/channel_id',
-        value: 'C0123456789',
-        team_id: 'T0123456789',
-        score: 75,
-        enterprise_id: 'E0123456789',
-      },
-    ],
-  },
-};
-expectAssignable<GenericMessageEvent>(messageWithAppContext);
